# resources/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Toy Domain Service API",
    "description": "도메인 핵심 로직을 제공하는 Domain Service API입니다.\n\n## 인증 아키텍처\n**JWT 검증은 BFF/Aggregator에서 수행**하며, Domain Service는 `X-User-Id` 헤더로 사용자 ID만 전달받습니다.\n\n```\nClient → BFF (JWT 검증) → Domain Service (X-User-Id 헤더)\n```\n\n인증이 필요한 엔드포인트는 `X-User-Id` 헤더를 필수로 요구합니다.\n\n## Auth 도메인\n소셜 로그인 및 토큰 관리를 담당하는 원자적 도메인 서비스입니다.\n- 소셜 로그인 (GitHub, Naver, Kakao)\n- JWT 토큰 발급/갱신\n- 다중 소셜 계정 연동\n- Refresh Token Rotation\n\n## Timer 도메인\n목표 시점까지의 남은 시간을 관리하는 원자적 도메인 서비스입니다.\n- 타이머 생성/조회/수정/삭제 (CRUD + Upsert)\n- 목표 시간(target_at)까지의 남은 초(remaining_seconds) 계산\n- Soft Delete 지원\n\n### remaining_seconds 해석\n- 음수 (-3600): 목표까지 3600초 남음\n- 0: 목표 시점 도달\n- 양수 (+3600): 목표 시점 3600초 경과\n\n## File 도메인\nMinIO 기반의 파일 저장 및 관리를 담당하는 원자적 도메인 서비스입니다.\n- 파일 업로드/다운로드/삭제\n- Public/Private 가시성 관리\n- 임시 URL 생성 (Private 파일용)\n- Soft Delete 지원\n\n## Notification 도메인\n다채널 알림 발송을 담당하는 원자적 도메인 서비스입니다.\n- Email, SMS, Slack 채널 지원\n- 즉시발송(immediate), 예약발송(scheduled), 묶음발송(batched) 지원\n- 발송 대기열 및 로그 관리\n- 실패 시 재시도 기능\n\n## User Activity 도메인\nMongoDB 기반의 사용자 활동 로그를 관리하는 원자적 도메인 서비스입니다.\n- 활동 로그 CRUD (생성/조회/수정/삭제)\n- 사용자별 활동 조회 및 통계\n- 유연한 메타데이터 저장\n- 날짜/액션 기반 필터링",
    "version": "1.0.0",
    "contact": {
      "name": "Toy Domain Service"
    }
  },
  "servers": [
    {
      "url": "https://dev-core.shaul.link",
      "description": "Development Server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local Server"
    }
...
  "paths": {
...
    "/api/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "JWT 토큰 갱신",
        "description": "Refresh Token으로 새 Access Token과 Refresh Token을 발급받습니다.\n\n**Token Rotation:**\n- 갱신 시마다 새로운 Refresh Token 발급\n- 기존 Refresh Token은 즉시 무효화\n- 토큰 재사용 감지 시 해당 Token Family 전체 무효화",
        "operationId": "refreshToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "토큰 갱신 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh Token 만료/무효",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "로그아웃",
        "description": "현재 세션을 로그아웃합니다. Refresh Token을 무효화합니다. (Access Token은 BFF에서 관리)",
        "operationId": "logout",
        "parameters": [
          {"$ref": "#/components/parameters/X-User-Id"}
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그아웃 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "X-User-Id 헤더 누락",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout-all": {
      "post": {
        "tags": ["Auth"],
        "summary": "모든 디바이스에서 로그아웃",
        "description": "모든 디바이스/세션에서 로그아웃합니다. 사용자의 token_version을 증가시켜 기존 발급된 모든 토큰을 무효화합니다.",
        "operationId": "logoutAll",
        "parameters": [
          {"$ref": "#/components/parameters/X-User-Id"}
        ],
        "responses": {
          "200": {
            "description": "전체 로그아웃 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "X-User-Id 헤더 누락",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "현재 사용자 정보",
        "description": "현재 로그인된 사용자의 정보를 조회합니다.",
        "operationId": "getCurrentUser",
        "parameters": [
          {"$ref": "#/components/parameters/X-User-Id"}
        ],
        "responses": {
          "200": {
            "description": "사용자 정보 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "X-User-Id 헤더 누락",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "사용자 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/validate": {
      "post": {
        "tags": ["Auth"],
        "summary": "토큰 유효성 검증",
        "description": "JWT 토큰의 유효성을 검증합니다. (디버그/테스트용)\n\nAuthorization 헤더에 Bearer 토큰을 포함하여 요청합니다.",
        "operationId": "validateToken",
        "responses": {
          "200": {
            "description": "토큰 유효",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenValidationResponse"
                }
              }
            }
          },
          "401": {
            "description": "토큰 무효/만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "X-User-Id": {
        "name": "X-User-Id",
        "in": "header",
        "required": true,
        "description": "BFF/Aggregator에서 JWT 검증 후 전달하는 사용자 ID",
        "schema": {
          "type": "integer",
          "example": 1
        }
      }
    },
    "schemas": {
...
      "AuthTokenResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "access_token": {
                "type": "string",
                "description": "JWT Access Token (1시간)",
                "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
              },
              "refresh_token": {
                "type": "string",
                "description": "Refresh Token (7일)",
                "example": "550e8400-e29b-41d4-a716-446655440000"
              },
              "token_type": {
                "type": "string",
                "example": "bearer"
              },
              "expires_in": {
                "type": "integer",
                "description": "Access Token 만료 시간 (초)",
                "example": 3600
              },
              "user": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "홍길동"
                  },
                  "email": {
                    "type": "string",
                    "example": "user@example.com"
                  },
                  "avatar": {
                    "type": "string",
                    "nullable": true,
                    "example": "https://avatars.githubusercontent.com/..."
                  }
                }
              }
            },
            "required": ["access_token", "refresh_token", "token_type", "expires_in", "user"]
          }
        },
        "required": ["success", "data"]
      },
...
      "RefreshTokenRequest": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "Refresh Token",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        },
        "required": ["refresh_token"]
      },
      "LogoutRequest": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "Refresh Token (선택, 함께 무효화)",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "로그아웃되었습니다."
              }
            },
            "required": ["message"]
          }
        },
        "required": ["success", "data"]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              },
              "name": {
                "type": "string",
                "example": "홍길동"
              },
              "email": {
                "type": "string",
                "example": "user@example.com"
              },
              "avatar": {
                "type": "string",
                "nullable": true,
                "example": "https://avatars.githubusercontent.com/..."
              },
              "email_verified_at": {
                "type": "string",
                "format": "date-time",
                "nullable": true,
                "example": "2025-12-11T10:00:00+09:00"
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "example": "2025-12-01T09:00:00+09:00"
              },
              "social_accounts": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "provider": {
                      "type": "string",
                      "example": "github"
                    },
                    "linked_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-12-01T09:00:00+09:00"
                    }
                  }
                }
              }
            },
            "required": ["id", "name", "email", "created_at"]
          }
        },
        "required": ["success", "data"]
      },
      "TokenValidationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "valid": {
                "type": "boolean",
                "example": true
              },
              "user_id": {
                "type": "integer",
                "example": 1
              }
            },
            "required": ["valid", "user_id"]
          }
        },
        "required": ["success", "data"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "에러 코드",
                "example": "NOT_FOUND"
              },
              "message": {
                "type": "string",
                "description": "에러 메시지",
                "example": "Timer를 찾을 수 없습니다: my-timer"
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["success", "error"]
      },
...
    }
  }
}